2023_assignment1_MTM/requirements.txt
mysql-connector-python
setuptools
